LPD-45149 Move layout page template test utility classes to layout-page-template-test-util
