+++ modules/apps/layout/layout-page-template-test-util/src/main/java/com/liferay/layout/page/template/test/util/LayoutPageTemplateTestUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.page.template.service.test.util;
+package com.liferay.layout.page.template.test.util;
 
 import com.liferay.layout.page.template.constants.LayoutPageTemplateCollectionTypeConstants;
 import com.liferay.layout.page.template.constants.LayoutPageTemplateConstants;
